+++ product_reference/tests/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import test_base_reference
+from . import test_reference_uppercase